Release v1.1.0 (#3)
* Update documentation for upcoming v1.1

* Add the ability to specify custom ALB healthy HTTP codes

* exclude v1.0 from search results

* Remove jekyll-admin for security reasons

jekyll-admin's second-level dependency `rack` was flagged by GitHub's
dependabot as needing an upgrade for security reasons. Because we
don't use jekyll-admin, we are resolving the security issue by
removing it in entirety.

* Move terraform.tf to tutorial examples to avoid conflict

The `terraform.tf` file provded by Alec in this repo's PR #1
was an example, but is being interpreted by Terraform as a
real file to use.

We are going to move it out of the way to make sure that doesn't
happen.

* Add uninstallation instructions

* Fix a typo in the docs for the databases

* Reformat and add to debugging docs

* Copy v1.0 docs changes to v1.1

* Fix header in debugging docs

* Fix hosted Sentry installation

We were having issues logging to Sentry in before. These changes upgrade
Provose to the latest Sentry version and alter the setup to make it
work.

* Update containers when existing secrets change

Previously, if we created a secret and changed its value, the
change in value did not cause Terraform to update Elastic Container
Service.

By referencing the `aws_secretsmanager_secret_version` resource instead
of the `aws_secretsmanager_secret` resource, Terraform understands
that changes in existing secret values should force an update
in the Elastic Container Service.

* Reformat the code for ECS task execution role

It is cleaner to describe the role policy in Terraform HCL
and then convert it to JSON.

* Set proxy configuration for ECS EC2 hosts

Just in case the user sets an HTTP proxy, we should specify which
URLs we do not want to connect to via the proxy.

* Fix indentation for `containers` configuration

* Add CloudWatch permissions to Grafana configuration

Grafana can pull information from CloudWatch, but only if the
underlying instance has access to CloudWatch in its IAM
instance profile.

This sets up the Grafana instance to pull from CloudWatch.

* Added support for AWS Fargate Spot

* Update changelog

Co-authored-by: James Mishra <james@neocrym.com>